requirements.txt
alembic==1.7.5
attrs==22.1.0
blinker==1.4
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.19.2
python-slugify==5.0.2
SQLAlchemy==1.4.29
text-unidecode==1.3
tomli==2.0.1
Werkzeug==2.0.2
WTForms==3.0.1
zipp==3.7.0

marshmallow~=3.17.1